Use clap for argument parsing.
The various binaries are consolidated into a single `outpack` command,
and their command line arguments are unified through the use of the clap
crate.

Some of the error types around the crate are cleaned up a bit to allow
this consolidation. In particular the api and init modules now use
an `anyhow` error, and the query module uses a `thiserror` to keep a bit
more precision. There are still a lot of (ab)uses of `io::Error` around,
but these can be updated later.